backend/pyproject.toml
[tool.poetry]
name = "Template project"
version = "0.1.0"
description = "template project"
license = "MIT"
authors=["Denis Pavlov"]


[tool.poetry.dependencies]
python = "3.10.*"
fastapi = {extras = ["all"], version = "0.110.*"}
uvicorn = "0.29.*"
itsdangerous = "2.1.2"
pydantic-settings = "2.2.1"
alembic = "1.13.*"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.16"
fastapi-debug-toolbar = "^0.6.2"
pytest = "^8.1.1"
sqlalchemy-utils = "^0.41.2"
pytest-asyncio = "^0.23.6"
factory-boy = "^3.2.0"
boto3 = "^1.34.122"
fastapi-filter = "^2.0.0"
redis = "^5.0.8"
celery = "^5.4.0"

# Linting
ruff = "0.0.*"
async-factory-boy = "^1.0.1"


[tool.ruff]
select = ["ALL"]
ignore = [
    "A002",
    "ANN201",
    "A003",
    "EXE002",
    "D100",
    "D104",
    "RUF002",
    "FBT001",
    "FBT002",
    "B008",
    "BLE001",
]
exclude = [
    "**/migrations/**",
]
line-length = 100
